requirements.txt
langchain 
ipykernel
python-dotenv
langchain-community
pypdf
bs4
langchain-text-splitters
langchain-openai
sentence-transformers
langchain-huggingface
faiss-cpu
langchain-anthropic
langchain_chroma
beautifulsoup4
langchain_groq
fastapi
uvicorn 
langserve 